Allow pasting ZED urls in the command palette in development
